+++ server/controller/folder.controller.js
@@ -1,9 +1,12 @@
 const db = require("../models");
 const Folder = db.folder
 const Page = db.page
+const guardToken = require("../middleware/guardToken")
 
 // Create a new page
 exports.create = (req, res) => {
+    if(guardToken.guardToken(req,res)) return  false
+
     const folder = new Folder({
         name: req.body.name,
     });
@@ -65,6 +68,8 @@ exports.findById = (req, res) => {
 
 // Update a folder by the id in the request
 exports.update = (req, res) => {
+    if(guardToken.guardToken(req,res)) return  false
+
     if (!req.body) {
         return res.status(400).send({
             message: "Data to update can not be empty!"
@@ -90,8 +95,9 @@ exports.update = (req, res) => {
 
 // Delete a folder by id
 exports.delete = (req, res) => {
+    if(guardToken.guardToken(req,res)) return  false
+    
     const id = req.params.id;
-    console.log('folder id', id)
     Page.deleteMany({ folder: id }).then(() => {
         Folder.findByIdAndRemove(id)
             .then(data => {
