+++ server/controller/page.controller.js
@@ -1,10 +1,11 @@
 const db = require("../models");
 const Page = db.page
-
+const guardToken = require("../middleware/guardToken")
 
 // Create a new page
 exports.create = (req, res) => {
-    console.log("req", req.body);
+    if(guardToken.guardToken(req,res)) return  false
+
     const page = new Page({
         name: req.body.name,
         folder: req.body.folder,
@@ -55,6 +56,8 @@ exports.findById = (req, res) => {
 
 //delete pdf file 
 exports.deletePdf = (req, res) => {
+    if(guardToken.guardToken(req,res)) return  false
+
     if (!req.body) {
         return res.status(400).send({
             message: "Data to update can not be empty!"
@@ -64,6 +67,7 @@ exports.deletePdf = (req, res) => {
     const id = req.params.id;
 
     Page.updateOne(
+        
         { _id: id },
         {
             $pull: { files: { id: req.body.idPdf }}
@@ -84,6 +88,8 @@ exports.deletePdf = (req, res) => {
 
 // Update a page by the id in the request
 exports.update = (req, res) => {
+    if(guardToken.guardToken(req,res)) return  false
+
     if (!req.body) {
         return res.status(400).send({
             message: "Data to update can not be empty!"
@@ -118,6 +124,8 @@ exports.update = (req, res) => {
 
 // Delete a page by id
 exports.delete = (req, res) => {
+    if(guardToken.guardToken(req,res)) return  false
+
     const id = req.params.id;
 
     Page.findByIdAndRemove(id)
@@ -141,7 +149,7 @@ exports.delete = (req, res) => {
 };
 
 exports.findPage = (req, res) => {
-    console.log("page")
+    
     Page.find({
         name:{ $regex : new RegExp(`${req.body.title}`, "i") }
     })
